refactor Milestone and Step components to be reusable across osc and ssj
